Restructure pages to improve component reusability
